Slightly reorganize const eval code
